+++ .template-lintrc.js
@@ -5,7 +5,10 @@ module.exports = {
   extends: 'recommended',
 
   rules: {
-    'no-bare-strings': false,
+    'no-bare-strings': true,
     'quotes': 'double'
-  }
+  },
+  ignore: [
+    '**/tests/dummy/**'
+  ]
 };