+++ packages/@nucleus/tests/dummy/app/components/nucleus-modal/demo-3.js
@@ -1,10 +1,13 @@
 // BEGIN-SNIPPET slider-component.js
 import Component from '@ember/component';
+import { inject } from '@ember/service';
+import { get } from '@ember/object'
 
 export default Component.extend({
+  flashMessages: inject(),
   actions: {
     onSubmit(value) {
-      alert(value);
+      get(this, 'flashMessages').success(value);
     }
   }
 });