+++ packages/modal/addon/constants/nucleus-modal.js
@@ -1,0 +1,3 @@
+const focusableElements = 'a[href]:not([disabled]), button:not([disabled]), textarea:not([disabled]), input[type="text"]:not([disabled]), input[type="radio"]:not([disabled]), input[type="checkbox"]:not([disabled]), select:not([disabled])';
+
+export { focusableElements };