+++ packages/toast-message/config/environment.js
@@ -1,5 +1,6 @@
 'use strict';
 
+/* eslint-env node */
 module.exports = function(environment) {
   let ENV = {
     modulePrefix: 'nucleus',