+++ packages/@nucleus/CHANGELOG.md
@@ -3,6 +3,125 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://github.com/freshdesk/nucleus/compare/nucleus@0.5.5...nucleus@0.7.0) (2020-03-20)
+
+
+### Bug Fixes
+
+* **Inline-banner:** refactored yield ([#151](https://github.com/freshdesk/nucleus/issues/151)) ([0ce4615](https://github.com/freshdesk/nucleus/commit/0ce46152619aae09c5363aad06499d30c86d5336))
+
+
+### Features
+
+* **tabs:** Implementation ([#138](https://github.com/freshdesk/nucleus/issues/138)) ([598e3e6](https://github.com/freshdesk/nucleus/commit/598e3e6e5a18efa8d8127c7c4165629c8e44cc92))
+
+
+
+
+
+# [0.6.0](https://github.com/freshdesk/nucleus/compare/nucleus@0.5.5...nucleus@0.6.0) (2020-03-20)
+
+
+### Bug Fixes
+
+* **Inline-banner:** refactored yield ([#151](https://github.com/freshdesk/nucleus/issues/151)) ([0ce4615](https://github.com/freshdesk/nucleus/commit/0ce46152619aae09c5363aad06499d30c86d5336))
+
+
+### Features
+
+* **tabs:** Implementation ([#138](https://github.com/freshdesk/nucleus/issues/138)) ([598e3e6](https://github.com/freshdesk/nucleus/commit/598e3e6e5a18efa8d8127c7c4165629c8e44cc92))
+
+
+
+
+
+## [0.5.5](https://github.com/freshdesk/nucleus/compare/nucleus@0.5.4...nucleus@0.5.5) (2020-03-12)
+
+
+### Bug Fixes
+
+* **modal:** Fixes bugs for focus-trap and autofocus ([#148](https://github.com/freshdesk/nucleus/issues/148)) ([198ec6f](https://github.com/freshdesk/nucleus/commit/198ec6f25ddec0bcff6f5e97a38c507b2b181a0d))
+
+
+
+
+
+## 0.5.4 (2020-03-09)
+
+**Note:** Version bump only for package nucleus
+
+
+
+
+
+## 0.5.3 (2020-02-25)
+
+**Note:** Version bump only for package nucleus
+
+
+
+
+
+## 0.5.2 (2020-02-25)
+
+**Note:** Version bump only for package nucleus
+
+
+
+
+
+## 0.5.1 (2020-02-19)
+
+**Note:** Version bump only for package nucleus
+
+
+
+
+
+# 0.5.0 (2020-02-18)
+
+
+### Bug Fixes
+
+* **button:** CSS code cleanup and enhancements ([#113](https://github.com/freshdesk/nucleus/issues/113)) ([2ea58af](https://github.com/freshdesk/nucleus/commit/2ea58afc6d1d68d685d87ecb8408b033885d63c2))
+
+
+### Features
+
+* **icon:** Implemented {{nucleus-icon}} ([#85](https://github.com/freshdesk/nucleus/issues/85)) ([a567468](https://github.com/freshdesk/nucleus/commit/a5674681147082f5f9790738cec1d632cac5debe))
+* **icon:** variant property ([#118](https://github.com/freshdesk/nucleus/issues/118)) ([7905cac](https://github.com/freshdesk/nucleus/commit/7905cac643897074a1276e79dc10f35d6a3003c4))
+
+
+
+
+
+## [0.4.5](https://github.com/freshdesk/nucleus/compare/nucleus@0.4.4...nucleus@0.4.5) (2020-02-14)
+
+
+### Bug Fixes
+
+* **button:** CSS code cleanup and enhancements ([#113](https://github.com/freshdesk/nucleus/issues/113)) ([2ea58af](https://github.com/freshdesk/nucleus/commit/2ea58afc6d1d68d685d87ecb8408b033885d63c2))
+
+
+
+
+
+## [0.4.4](https://github.com/freshdesk/nucleus/compare/nucleus@0.4.3...nucleus@0.4.4) (2020-02-10)
+
+**Note:** Version bump only for package nucleus
+
+
+
+
+
+## [0.4.3](https://github.com/freshdesk/nucleus/compare/nucleus@0.4.2...nucleus@0.4.3) (2020-02-07)
+
+**Note:** Version bump only for package nucleus
+
+
+
+
+
 # [0.2.0](https://github.com/freshdesk/nucleus/compare/nucleus@0.1.8...nucleus@0.2.0) (2020-01-09)